n8n_minimal/requirements.txt
fastapi
uvicorn
python-dotenv
motor
pymongo
cryptography
python-jose
passlib
httpx
apscheduler
sse-starlette
prometheus-fastapi-instrumentator
RestrictedPython
langchain
langchain-openai
langchain-anthropic
langchain-groq
langgraph
bs4
chroma 